+++ packages/console-viewer/src/pages/ApplicationVersionPage.tsx
@@ -1,0 +1,1 @@
+export { ApplicationVersionPage as default } from "../ui.js";